+++ src/app/auth/@authModal/(.)after-register/page.tsx
@@ -1,28 +1,14 @@
 'use client';
 
-import { useRouter, useSearchParams } from 'next/navigation';
+import { Suspense } from 'react';
 
-import { ROUTES } from '@common/constants/routes.constants';
-
-import ModalLayoutWithoutPortal from '@common/components/layouts/ModalLayoutWithoutPortal/ModalLayoutWithoutPortal.client';
-import CongratulateSignUpModal from '@common/components/modals/CongratulateSignUpModal/CongratulateSignUpModal.client';
+import InnerAfterRegisterModal from '@features/auth/components/InnerAfterRegisterModal/InnerAfterRegisterModal.client';
 
 const AfterRegisterModal = () => {
-  const router = useRouter();
-  const handleClose = () => router.back();
-  const handleCreateWorkspace = () => router.push(ROUTES.ONBOARDING);
-
-  const query = useSearchParams();
-  const isInvitedRegister = query.get('invited') === 'true';
-
   return (
-    <ModalLayoutWithoutPortal>
-      <CongratulateSignUpModal
-        onClose={handleClose}
-        onWorkspaceCreate={handleCreateWorkspace}
-        isInvited={isInvitedRegister}
-      />
-    </ModalLayoutWithoutPortal>
+    <Suspense fallback={<div>Loading...</div>}>
+      <InnerAfterRegisterModal />
+    </Suspense>
   );
 };
